update-3 collision

+++ hash_tables/hash_tables.h
@@ -46,7 +46,7 @@ typedef struct hash_table_s
 	 unsigned long int size;
 	 hash_node_t **array;
 	 LinkedList_t **overflow_buckets;
-	 int count;
+	 unsigned long int  count;
 } hash_table_t;
 
 hash_node_t *create_item(const char *key, const char * value);
